Added separate `ollama` provider
Created a separate file `ollama.py` as a unique provider. Refactored other code accordingly.

Also changed the `Ollama` class to `ChatOllama` so that it can support binding tools to the LLM.

Updated the imports to come from `langchain_ollama` instead of `langchain_community`

Tested on several Ollama models, both LLMs and embedding models: `mxbai-embed-large`, `nomic-embed-text`, `ima/deepseek-math`, `mathstral`, `qwen2-math`, `snowflake-arctic-embed`, `mistral`, `llama3.1`, `starcoder2:15b-instruct`
