+++ packages/jupyter-ai-magics/jupyter_ai_magics/embedding_providers.py
@@ -10,7 +10,6 @@ from langchain.pydantic_v1 import BaseModel, Extra
 from langchain_community.embeddings import (
     GPT4AllEmbeddings,
     HuggingFaceHubEmbeddings,
-    OllamaEmbeddings,
     QianfanEmbeddingsEndpoint,
 )
 
@@ -63,19 +62,6 @@ class BaseEmbeddingsProvider(BaseModel):
             model_kwargs[self.__class__.model_id_key] = kwargs["model_id"]
 
         super().__init__(*args, **kwargs, **model_kwargs)
-
-
-class OllamaEmbeddingsProvider(BaseEmbeddingsProvider, OllamaEmbeddings):
-    id = "ollama"
-    name = "Ollama"
-    # source: https://ollama.com/library
-    models = [
-        "nomic-embed-text",
-        "mxbai-embed-large",
-        "all-minilm",
-        "snowflake-arctic-embed",
-    ]
-    model_id_key = "model"
 
 
 class HfHubEmbeddingsProvider(BaseEmbeddingsProvider, HuggingFaceHubEmbeddings):
