+++ packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
@@ -33,7 +33,6 @@ from langchain_community.llms import (
     AI21,
     GPT4All,
     HuggingFaceEndpoint,
-    Ollama,
     Together,
 )
 from langchain_core.language_models.chat_models import BaseChatModel
@@ -706,20 +705,6 @@ class HfHubProvider(BaseProvider, HuggingFaceEndpoint):
     async def _acall(self, *args, **kwargs) -> Coroutine[Any, Any, str]:
         return await self._call_in_executor(*args, **kwargs)
 
-
-class OllamaProvider(BaseProvider, Ollama):
-    id = "ollama"
-    name = "Ollama"
-    model_id_key = "model"
-    help = (
-        "See [https://www.ollama.com/library](https://www.ollama.com/library) for a list of models. "
-        "Pass a model's name; for example, `deepseek-coder-v2`."
-    )
-    models = ["*"]
-    registry = True
-    fields = [
-        TextField(key="base_url", label="Base API URL (optional)", format="text"),
-    ]
 
 
 class TogetherAIProvider(BaseProvider, Together):
